[Siddhartha] modified: search by city or state condition

+++ src/main/java/com/cg/adp/AddressBookExecutor.java
@@ -3,6 +3,7 @@ package com.cg.adp;
 import java.util.Scanner;
 
 import com.cg.adp.service.AddressBookService;
+import com.cg.adp.service.PersonContactService;
 
 public class AddressBookExecutor {
 	private Scanner sc;
@@ -20,40 +21,58 @@ public class AddressBookExecutor {
 		boolean compuKeeper = true;
 		while (compuKeeper) {
 			System.out.println("1.Add address book.");
-			System.out.println("2.Exit.");
+			System.out.println("2.Search by city");
+			System.out.println("3.Search by state");
+			System.out.println("4.Exit.");
 			int op = abe.sc.nextInt();
 			switch (op) {
 			case 1:
 				abService.addAddressBook();
 				break;
 			case 2:
-				compuKeeper = false;
-				break;
-			}
-		}
-		boolean computation = true;
-		while (computation) {
-			System.out.println("1.Add contact. ");
-			System.out.println("2.Edit Contact.");
-			System.out.println("3.Remove Contact");
-			System.out.println("4.Exit.");
-			int option = abe.sc.nextInt();
-
-			switch (option) {
-			case 1:
-				abService.addContact();
-				break;
-			case 2:
-				abService.updateAddressBook();
+				PersonContactService pcService = new PersonContactService();
+				abService.getAbks().searchByCity(abService.getAbe(), pcService.gettingFirstName());
 				break;
 			case 3:
-				abService.removeContact();
+				pcService = new PersonContactService();
+				abService.getAbks().searchByState(abService.getAbe(), pcService.gettingFirstName());
 				break;
+
 			case 4:
-				computation = false;
+				compuKeeper = false;
+				System.out.println("THANKS :)");
 				break;
 			}
+			System.out.println("Enter the address book you want to work on:");
+			String abName = abe.sc.next();
+			if (abService.getAbks().searchAddressBook(abService.getAbe(), abName)) {
 
+				boolean computation = true;
+				while (computation) {
+					System.out.println("1.Add contact. ");
+					System.out.println("2.Edit Contact.");
+					System.out.println("3.Remove Contact");
+					System.out.println("4.Exit.");
+					int option = abe.sc.nextInt();
+
+					switch (option) {
+					case 1:
+						abService.addContact();
+						break;
+					case 2:
+						abService.updateAddressBook();
+						break;
+					case 3:
+						abService.removeContact();
+						break;
+					case 4:
+						computation = false;
+						System.out.println("THANKS :)");
+						break;
+					}
+
+				}
+			}
 		}
 
 	}
